(toString(int)) : Fix a bug.
git-svn-id: http://es-operating-system.googlecode.com/svn/trunk/www@2231 d15c3b8f-5b43-0410-be93-4d5a55ce2954

+++ utf.h
@@ -221,7 +221,7 @@ inline std::u16string toString(unsigned int value)
 inline std::u16string toString(int value)
 {
     if (value < 0)
-        return u"-" + toString(-value);
+        return u"-" + toString(static_cast<unsigned int>(-value));
     else
         return toString(static_cast<unsigned int>(value));
 }
